+++ FirebaseDatabase/Sources/Api/FIRDatabase.m
@@ -153,13 +153,15 @@
     }
     FParsedUrl *parsedUrl = [FUtilities parseUrl:databaseUrl];
     [FValidation validateFrom:@"referenceFromURL:" validURL:parsedUrl];
-    if (![parsedUrl.repoInfo.host isEqualToString:_repoInfo.host]) {
-        [NSException
-             raise:@"InvalidDatabaseURL"
-            format:
-                @"Invalid URL (%@) passed to getReference(). URL was expected "
-                 "to match configured Database URL: %@",
-                databaseUrl, [self reference].URL];
+
+    BOOL isInvalidHost =
+        !parsedUrl.repoInfo.isCustomHost &&
+        ![_repoInfo.host isEqualToString:parsedUrl.repoInfo.host];
+    if (isInvalidHost) {
+        [NSException raise:@"InvalidDatabaseURL"
+                    format:@"Invalid URL (%@) passed to getReference(). URL "
+                           @"was expected to match configured Database URL: %@",
+                           databaseUrl, _repoInfo.host];
     }
     return [[FIRDatabaseReference alloc] initWithRepo:self.repo
                                                  path:parsedUrl.path];
@@ -171,6 +173,25 @@
     dispatch_async([FIRDatabaseQuery sharedQueue], ^{
       [self.repo purgeOutstandingWrites];
     });
+}
+
+- (void)useEmulatorWithHost:(NSString *)host port:(NSInteger)port {
+    if (host.length == 0) {
+        [NSException raise:NSInvalidArgumentException
+                    format:@"Cannot connect to nil or empty host."];
+    }
+    if (self.repo != nil) {
+        [NSException
+             raise:NSInternalInconsistencyException
+            format:@"Cannot connect to emulator after database initialization. "
+                   @"Call useEmulator(host:port:) before creating a database "
+                   @"reference or trying to load data."];
+    }
+    NSString *fullHost =
+        [NSString stringWithFormat:@"%@:%li", host, (long)port];
+    FRepoInfo *emulatorInfo = [[FRepoInfo alloc] initWithInfo:self.repoInfo
+                                                 emulatedHost:fullHost];
+    self->_repoInfo = emulatorInfo;
 }
 
 - (void)goOnline {