+++ Functions/FirebaseFunctions/Public/FirebaseFunctions/FIRFunctions.h
@@ -25,6 +25,12 @@ NS_ASSUME_NONNULL_BEGIN
 NS_SWIFT_NAME(Functions)
 @interface FIRFunctions : NSObject
 
+/**
+ * The current emulator origin, or nil if it is not set.
+ */
+@property(nonatomic, readonly, nullable) NSString *emulatorOrigin;
+
+/** :nodoc: */
 - (id)init NS_UNAVAILABLE;
 
 /**
@@ -77,9 +83,19 @@ NS_SWIFT_NAME(Functions)
 /**
  * Changes this instance to point to a Cloud Functions emulator running locally.
  * See https://firebase.google.com/docs/functions/local-emulator
- * @param origin The origin of the local emulator, such as "http://localhost:5005".
+ * @param origin The origin of the local emulator, such as "localhost:5005".
  */
-- (void)useFunctionsEmulatorOrigin:(NSString *)origin NS_SWIFT_NAME(useFunctionsEmulator(origin:));
+- (void)useFunctionsEmulatorOrigin:(NSString *)origin
+    NS_SWIFT_NAME(useFunctionsEmulator(origin:))
+        __attribute__((deprecated("Use useEmulator(host:port:) instead.")));
+
+/**
+ * Changes this instance to point to a Cloud Functions emulator running locally.
+ * See https://firebase.google.com/docs/functions/local-emulator
+ * @param host The host of the local emulator, such as "localhost".
+ * @param port The port of the local emulator, for example 5005.
+ */
+- (void)useEmulatorWithHost:(NSString *)host port:(NSInteger)port;
 
 @end