+++ Functions/CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+- [added] Made emulator connection API consistent between Auth, Database, Firestore, and Functions (#5916).
+
 # v7.1.0
 - [added] Added a constructor to set a custom domain. (#6787)